feat: addition/subtraction/multiplication/division support

+++ database/migrations/2022_12_19_213706_create_calculations_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('calculations', function (Blueprint $table) {
             $table->id();
             $table->text('expression');
-            $table->float('result');
+            $table->double('result', 12, 3);
             $table->timestamps();
         });
     }
